# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LocalOS self-hosted AI operating system API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: status
    description: System status
  - name: projects
    description: Project management
  - name: chats
    description: Chat management
  - name: files
    description: File management
  - name: knowledge
    description: Knowledge base
  - name: models
    description: Model management
  - name: filesystem
    description: Filesystem operations
  - name: chat
    description: AI chat streaming
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /status:
    get:
      operationId: getSystemStatus
      tags: [status]
      summary: Get system status
      description: Returns status of all LocalOS subsystems
      responses:
        "200":
          description: System status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemStatus"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects/{projectId}/chats:
    get:
      operationId: listChats
      tags: [chats]
      summary: List chats for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of chats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chat"
    post:
      operationId: createChat
      tags: [chats]
      summary: Create a new chat
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "201":
          description: Created chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"

  /projects/{projectId}/chats/{chatId}:
    get:
      operationId: getChat
      tags: [chats]
      summary: Get a chat with messages
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatWithMessages"
        "404":
          description: Not found
    patch:
      operationId: updateChat
      tags: [chats]
      summary: Update a chat
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatUpdate"
      responses:
        "200":
          description: Updated chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
    delete:
      operationId: deleteChat
      tags: [chats]
      summary: Delete a chat
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects/{projectId}/chats/{chatId}/messages:
    post:
      operationId: addMessage
      tags: [chats]
      summary: Add a message to a chat
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Added message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /projects/{projectId}/files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List files for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectFile"
    post:
      operationId: createFile
      tags: [files]
      summary: Create a file in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectFileInput"
      responses:
        "201":
          description: Created file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"

  /projects/{projectId}/files/{fileId}:
    get:
      operationId: getFile
      tags: [files]
      summary: Get a file
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"
        "404":
          description: Not found
    patch:
      operationId: updateFile
      tags: [files]
      summary: Update a file
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectFileUpdate"
      responses:
        "200":
          description: Updated file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects/{projectId}/knowledge:
    get:
      operationId: listKnowledge
      tags: [knowledge]
      summary: List knowledge docs for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of knowledge docs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeDoc"
    post:
      operationId: createKnowledge
      tags: [knowledge]
      summary: Add a knowledge document
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeDocInput"
      responses:
        "201":
          description: Created knowledge doc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeDoc"

  /projects/{projectId}/knowledge/{docId}:
    delete:
      operationId: deleteKnowledge
      tags: [knowledge]
      summary: Delete a knowledge document
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: docId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /models:
    get:
      operationId: listModels
      tags: [models]
      summary: List available models
      responses:
        "200":
          description: List of models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OllamaModel"

  /models/pull:
    post:
      operationId: pullModel
      tags: [models]
      summary: Pull (download) a model via Ollama
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelPullInput"
      responses:
        "200":
          description: Pull result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelPullResult"

  /models/{modelName}:
    delete:
      operationId: deleteModel
      tags: [models]
      summary: Delete a model
      parameters:
        - name: modelName
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /filesystem:
    get:
      operationId: listFilesystem
      tags: [filesystem]
      summary: List directory contents
      parameters:
        - name: path
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Directory listing
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FilesystemEntry"

  /filesystem/read:
    get:
      operationId: readFilesystemFile
      tags: [filesystem]
      summary: Read a file from the filesystem
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilesystemFileContent"

  /filesystem/write:
    post:
      operationId: writeFilesystemFile
      tags: [filesystem]
      summary: Write a file to the filesystem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilesystemWriteInput"
      responses:
        "200":
          description: Write result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilesystemWriteResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    SystemStatus:
      type: object
      required: [backend, database, ollama, storage, version]
      properties:
        backend:
          type: boolean
        database:
          type: boolean
        ollama:
          type: boolean
        storage:
          type: boolean
        version:
          type: string
        ollamaVersion:
          type: ["string", "null"]
        models:
          type: array
          items:
            type: string
        storageSize:
          type: ["string", "null"]

    Project:
      type: object
      required: [id, name, description, color, createdAt, updatedAt, chatCount, fileCount]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        color:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        chatCount:
          type: integer
        fileCount:
          type: integer

    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        color:
          type: string

    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        color:
          type: string

    Chat:
      type: object
      required: [id, projectId, title, messageCount, createdAt, updatedAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        messageCount:
          type: integer
        model:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    ChatWithMessages:
      type: object
      required: [id, projectId, title, messages, createdAt, updatedAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        model:
          type: ["string", "null"]
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        createdAt:
          type: string
        updatedAt:
          type: string

    ChatInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        model:
          type: string

    ChatUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        model:
          type: string

    Message:
      type: object
      required: [id, chatId, role, content, createdAt]
      properties:
        id:
          type: integer
        chatId:
          type: integer
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        createdAt:
          type: string

    MessageInput:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string

    ProjectFile:
      type: object
      required: [id, projectId, name, path, content, language, createdAt, updatedAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        path:
          type: string
        content:
          type: string
        language:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string

    ProjectFileInput:
      type: object
      required: [name, path]
      properties:
        name:
          type: string
          minLength: 1
        path:
          type: string
        content:
          type: string
        language:
          type: string

    ProjectFileUpdate:
      type: object
      properties:
        name:
          type: string
        content:
          type: string
        language:
          type: string

    KnowledgeDoc:
      type: object
      required: [id, projectId, name, content, type, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        content:
          type: string
        type:
          type: string
        size:
          type: ["integer", "null"]
        createdAt:
          type: string

    KnowledgeDocInput:
      type: object
      required: [name, content, type]
      properties:
        name:
          type: string
          minLength: 1
        content:
          type: string
        type:
          type: string

    OllamaModel:
      type: object
      required: [name, type, size, ram, vram, status]
      properties:
        name:
          type: string
        displayName:
          type: string
        type:
          type: string
          enum: [reasoning, coding, vision, embedding, search]
        size:
          type: string
        ram:
          type: string
        vram:
          type: string
        status:
          type: string
          enum: [not_installed, ready]
        progress:
          type: ["number", "null"]
        ollamaId:
          type: ["string", "null"]

    ModelPullInput:
      type: object
      required: [name]
      properties:
        name:
          type: string

    ModelPullResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string

    FilesystemEntry:
      type: object
      required: [name, path, type]
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory]
        size:
          type: ["integer", "null"]
        extension:
          type: ["string", "null"]

    FilesystemFileContent:
      type: object
      required: [path, content]
      properties:
        path:
          type: string
        content:
          type: string
        size:
          type: integer

    FilesystemWriteInput:
      type: object
      required: [path, content]
      properties:
        path:
          type: string
        content:
          type: string

    FilesystemWriteResult:
      type: object
      required: [success, path]
      properties:
        success:
          type: boolean
        path:
          type: string