# lib/api-zod/src/generated/types/knowledgeDoc.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LocalOS self-hosted AI operating system API
 * OpenAPI spec version: 0.1.0
 */

export interface KnowledgeDoc {
  id: number;
  projectId: number;
  name: string;
  content: string;
  type: string;
  /** @nullable */
  size?: number | null;
  createdAt: string;
}