# lib/api-zod/src/generated/types/modelPullResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LocalOS self-hosted AI operating system API
 * OpenAPI spec version: 0.1.0
 */

export interface ModelPullResult {
  success: boolean;
  message: string;
}